crd layout collpase/expand at once
